Improved to use RANGE TABLE with dynamic labels
Applied additional items covered in class where a separate Range Table and dynamic row labels were used. Additional changes to refine the boundaries will require minimal updates without changes to the formula - it only requires the business to update the Range Table. And if VBA script is used, it is one more step to easier maintenance.